Set the vboxnet1 interface in promiscuous mode to ping OpenStack VMs
Host is not able to ping OpenStack VMs. The reason why is that frames
destinated to other machines than the vagrant VM are dropped by VirtualBox.

Here is what is observed when ping is triggered from the host to an OpenStack
VM: tcpdump displays ARP requests (broadcast) and replies from enp0s9
interface of the VBox VM but ICMP requests are not received, while they
are displayed from the vboxnet1 interface of host. It shows that VirtualBox
drops frames between them.

After investigating VirtualBox configuration options, the solution appears to
activate the promiscuous mode of the VirtualBox NIC. This patch add the related
line in the Vagrantfile.

Mind that on my computer, running archlinux, the VirtualBox interfaces are not
removed when the VM is destroyed. It must be removed manually to be sure the
promiscuous mode is activated. For that, one can use:
`VBoxManage hostonlyif remove vboxnet1`

To check if the desired mode is activated, you can check the NIC' state:
1. Find out the name of the VM: `VBoxManage list vms`
2. Display VM info: `VBoxManage showvminfo <vm_name> | grep "Promisc Policy:"`